+++ doc/conf.py
@@ -53,20 +53,37 @@ extensions = [
 # Create links to references within kikuchipy's documentation to these
 # packages
 intersphinx_mapping = {
+    # Package
     "dask": ("https://docs.dask.org/en/stable", None),
     "diffpy.structure": ("https://www.diffpy.org/diffpy.structure", None),
     "diffsims": ("https://diffsims.readthedocs.io/en/latest", None),
     "hyperspy": ("https://hyperspy.org/hyperspy-doc/current", None),
+    "h5py": ("https://docs.h5py.org/en/stable", None),
+    "imageio": ("https://imageio.readthedocs.io/en/stable", None),
     "matplotlib": ("https://matplotlib.org/stable", None),
+    "numba": ("https://numba.pydata.org/numba-doc/latest", None),
     "numpy": ("https://numpy.org/doc/stable", None),
     "orix": ("https://orix.readthedocs.io/en/stable", None),
+    "pooch": ("https://www.fatiando.org/pooch/latest", None),
     "pyebsdindex": ("https://pyebsdindex.readthedocs.io/en/stable", None),
     "python": ("https://docs.python.org/3", None),
     "pyvista": ("https://docs.pyvista.org", None),
     "scipy": ("https://docs.scipy.org/doc/scipy", None),
     "skimage": ("https://scikit-image.org/docs/stable", None),
     "sklearn": ("https://scikit-learn.org/stable", None),
-    "h5py": ("https://docs.h5py.org/en/stable", None),
+    # Docs
+    "black": ("https://black.readthedocs.io/en/stable", None),
+    "conda": ("https://conda.io/projects/conda/en/latest", None),
+    "defdap": ("https://defdap.readthedocs.io/en/latest", None),
+    "nbsphinx": ("https://nbsphinx.readthedocs.io/en/latest", None),
+    "nbval": ("https://nbval.readthedocs.io/en/latest", None),
+    "numpydoc": ("https://numpydoc.readthedocs.io/en/latest", None),
+    "pythreejs": ("https://pythreejs.readthedocs.io/en/stable", None),
+    "pyxem": ("https://pyxem.readthedocs.io/en/latest", None),
+    "readthedocs": ("https://docs.readthedocs.io/en/stable", None),
+    "sphinx": ("https://www.sphinx-doc.org/en/master", None),
+    "sphinx-gallery": ("https://sphinx-gallery.github.io/stable", None),
+    "xcdskd": ("https://xcdskd.readthedocs.io/en/latest", None),
 }
 
 # Add any paths that contain templates here, relative to this directory.
@@ -77,14 +94,30 @@ templates_path = ["_templates"]
 # affects html_static_path and html_extra_path.
 exclude_patterns = ["build", "_static/logo/*.ipynb"]
 
-# HTML theme: furo
-# https://pradyunsg.me/furo/
-html_theme = "furo"
-
-# Add any paths that contain custom static files (such as style sheets)
-# here, relative to this directory. They are copied after the builtin
-# static files, so a file named "default.css" will overwrite the builtin
-# "default.css".
+# HTML theming: pydata-sphinx-theme
+# https://pydata-sphinx-theme.readthedocs.io
+html_theme = "pydata_sphinx_theme"
+html_theme_options = {
+    "github_url": "https://github.com/pyxem/kikuchipy",
+    "header_links_before_dropdown": 6,
+    "icon_links": [
+        {
+            "name": "Gitter",
+            "url": "https://gitter.im/pyxem/kikuchipy",
+            "icon": "fab fa-gitter",
+        },
+    ],
+    "logo": {"alt_text": "kikuchipy"},
+    "navigation_with_keys": False,
+    "show_toc_level": 2,
+    "use_edit_page_button": True,
+}
+html_context = {
+    "github_user": "pyxem",
+    "github_repo": "kikuchipy",
+    "github_version": "develop",
+    "doc_path": "doc",
+}
 html_static_path = ["_static"]
 html_css_files = ["custom.css"]
 
@@ -92,14 +125,17 @@ html_css_files = ["custom.css"]
 pygments_style = "friendly"
 
 # Logo
-html_logo = "_static/logo/plasma_logo.svg"
+html_logo = "_static/logo/plasma_banner.png"
 html_favicon = "_static/logo/plasma_favicon.png"
 
-# Whether to show all warnings when building the documentation
-nitpicky = True
-
-# -- nbsphinx
+# nbsphinx
+# --------
 # https://nbsphinx.readthedocs.io
+nbsphinx_execute = "auto"  # always, auto, never
+nbsphinx_allow_errors = True
+nbsphinx_execute_arguments = [
+    "--InlineBackend.rc=figure.facecolor='w'",
+]
 # Taken from nbsphinx' own nbsphinx configuration file, with slight
 # modifications to point nbviewer and Binder to the GitHub develop
 # branch links when the documentation is launched from a kikuchipy
@@ -108,59 +144,36 @@ if "dev" in version:
     release_version = "develop"
 else:
     release_version = "v" + version
-# This is processed by Jinja2 and inserted before each notebook
 nbsphinx_prolog = (
     r"""
 {% set docname = 'doc/' + env.doc2path(env.docname, base=None)[:-6] + '.ipynb' %}
 
-.. raw:: html
+.. admonition:: Live notebook
 
-    <style>a:hover { text-decoration: underline; }</style>
-
-    <div class="admonition note">
-      This page was generated from
-      <a class="reference external" href="https://github.com/pyxem/kikuchipy/blob/"""
+    You can run this notebook in a `live session <https://mybinder.org/v2/gh/pyxem/kikuchipy/"""
     + release_version
-    + r"""/{{ docname|e }}">{{ docname|e }}</a>.
-      Interactive online version:
-      <span style="white-space: nowrap;"><a href="https://mybinder.org/v2/gh/pyxem/kikuchipy/"""
+    + r"""?filepath={{ docname|e }}>`__ |Binder| or view it `on Github <https://github.com/pyxem/kikuchipy/blob/"""
     + release_version
-    + r"""?filepath={{ docname|e }}"><img alt="Binder badge" src="https://mybinder.org/badge_logo.svg" style="vertical-align:text-bottom"></a>.</span>
-      <script>
-        if (document.location.host) {
-          $(document.currentScript).replaceWith(
-            '<a class="reference external" ' +
-            'href="https://nbviewer.jupyter.org/url' +
-            (window.location.protocol == 'https:' ? 's/' : '/') +
-            window.location.host +
-            window.location.pathname.slice(0, -4) +
-            'ipynb">View in <em>nbviewer</em></a>.'
-          );
-        }
-      </script>
-    </div>
+    + r"""/{{ docname|e }}>`__.
 
-.. raw:: latex
-
-    \nbsphinxstartnotebook{\scriptsize\noindent\strut
-    \textcolor{gray}{The following section was generated from
-    \sphinxcode{\sphinxupquote{\strut {{ docname | escape_latex }}}} \dotfill}}
+.. |Binder| image:: https://static.mybinder.org/badge_logo.svg
+   :target: https://mybinder.org/v2/gh/pyxem/kikuchipy/"""
+    + release_version
+    + r"""?filepath={{ docname|e }}
 """
 )
-# https://nbsphinx.readthedocs.io/en/0.8.0/never-execute.html
-nbsphinx_execute = "auto"  # always, auto, never
-nbsphinx_allow_errors = True
-nbsphinx_execute_arguments = [
-    "--InlineBackend.rc=figure.facecolor='w'",
-    "--InlineBackend.rc=font.size=15",
-]
 
-# -- sphinxcontrib-bibtex
+# Whether to show all warnings when building the documentation
+nitpicky = False
+
+# sphinxcontrib-bibtex
+# --------------------
 # https://sphinxcontrib-bibtex.readthedocs.io
-bibtex_bibfiles = ["bibliography.bib"]
+bibtex_bibfiles = ["user/bibliography.bib"]
 bibtex_reference_style = "author_year"
 
-# -- sphinx-codeautolink
+# sphinx-codeautolink
+# -------------------
 codeautolink_custom_blocks = {
     "python3": None,
     "pycon3": "sphinx_codeautolink.clean_pycon",
@@ -234,9 +247,10 @@ def linkcode_resolve(domain, info):
         return None
 
 
-# -- PyVista
+# PyVista
+# -------
 # https://docs.pyvista.org
-pyvista.global_theme.window_size = [700, 700]
+pyvista.global_theme.window_size = [600, 600]
 pyvista.set_jupyter_backend("pythreejs")
 
 # -- Copy button customization (taken from PyVista)
@@ -244,7 +258,8 @@ pyvista.set_jupyter_backend("pythreejs")
 copybutton_prompt_text = r">>> ?|\.\.\. "
 copybutton_prompt_is_regexp = True
 
-# -- sphinx.ext.autodoc
+# sphinx.ext.autodoc
+# ------------------
 # https://www.sphinx-doc.org/en/master/usage/extensions/autodoc.html
 autosummary_ignore_module_all = False
 autosummary_imported_members = True
@@ -253,7 +268,8 @@ autodoc_default_options = {
     "show-inheritance": True,
 }
 
-# -- numpydoc
+# numpydoc
+# --------
 # https://numpydoc.readthedocs.io
 numpydoc_show_class_members = False
 numpydoc_use_plots = True
@@ -278,7 +294,8 @@ numpydoc_validation_checks = {
 }
 # fmt: on
 
-# -- matplotlib.sphinxext.plot_directive
+# matplotlib.sphinxext.plot_directive
+# -----------------------------------
 # https://matplotlib.org/stable/api/sphinxext_plot_directive_api.html
 plot_formats = ["png"]
 plot_html_show_source_link = False
@@ -308,8 +325,8 @@ def _str_examples(self):
 
 SphinxDocString._str_examples = _str_examples
 
-
-# -- Sphinx-Gallery
+# Sphinx-Gallery
+# --------------
 # https://sphinx-gallery.github.io
 sphinx_gallery_conf = {
     "backreferences_dir": "reference/generated",
@@ -318,12 +335,12 @@ sphinx_gallery_conf = {
     "filename_pattern": "^((?!sgskip).)*$",
     "gallery_dirs": "examples",
     "reference_url": {"kikuchipy": None},
-    "run_stale_examples": True,
+    "run_stale_examples": False,
     "show_memory": True,
 }
 autosummary_generate = True
 
 # Download example datasets prior to building the docs
-print("[kikuchipy] Downloading example datasets")
+print("[kikuchipy] Downloading example datasets (if not found in cache)")
 _ = kp.data.nickel_ebsd_large(allow_download=True)
 _ = kp.data.silicon_ebsd_moving_screen_in(allow_download=True)